FROMGIT: HID: i2c-hid: Reorganize so ACPI and OF are separate modules
This patch rejiggers the i2c-hid code so that the OF (Open Firmware
aka Device Tree) and ACPI support is separated out a bit.  The OF and
ACPI drivers are now separate modules that wrap the core module.

Essentially, what we're doing here:
* Make "power up" and "power down" a function that can be (optionally)
  implemented by a given user of the i2c-hid core.
* The OF and ACPI modules are drivers on their own, so they implement
  probe / remove / suspend / resume / shutdown.  The core code
  provides implementations that OF and ACPI can call into.

We'll organize this so that we now have 3 modules: the old i2c-hid
module becomes the "core" module and two new modules will depend on
it, handling probing the specific device.

As part of this work, we'll remove the i2c-hid "platform data"
concept since it's not needed.

Signed-off-by: Douglas Anderson <dianders@chromium.org>
Reviewed-by: Hans de Goede <hdegoede@redhat.com>
Signed-off-by: Benjamin Tissoires <benjamin.tissoires@redhat.com>
(cherry picked from commit b33752c300232d7f95dd9a4353947d0c9e6a0e52
 git://git.kernel.org/pub/scm/linux/kernel/git/hid/hid.git for-5.12/i2c-hid)

BUG=b:172594151
TEST=No changes

Change-Id: Ied4ce10d229cd7c69abf13a0361ba0b8d82eb9c5
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/2641077
Reviewed-by: Douglas Anderson <dianders@chromium.org>
Reviewed-by: Sean Paul <seanpaul@chromium.org>
Tested-by: Douglas Anderson <dianders@chromium.org>