FROMGIT: drm/panel-simple: Don't wait longer for HPD than hpd_absent_delay
If a panel has an hpd_absent_delay specified then we know exactly how
long the maximum time is before HPD must be asserted.  That means we
can use it as a timeout for polling the HPD pin instead of using an
arbitrary timeout.  This is especially useful for dealing with panels
that periodically fail to power on and need to be retried.  We can
detect the problem sooner.

Signed-off-by: Douglas Anderson <dianders@chromium.org>
Reviewed-by: Stephen Boyd <swboyd@chromium.org>
Reviewed-by: Linus Walleij <linus.walleij@linaro.org>
Signed-off-by: Linus Walleij <linus.walleij@linaro.org>
Link: https://patchwork.freedesktop.org/patch/msgid/20210115144345.v2.2.I183b1817610d7a82fdd3bc852e96d2985df9623f@changeid
(cherry picked from commit 6ec52621e77c81079e7d3858cca00344d4266a89
 https://github.com/freedesktop/drm-misc.git drm-misc-next)

BUG=b:176918319, b:172695850
TEST=Stress going between VT1 and VT2 w/ CL:2530131

Change-Id: I183b1817610d7a82fdd3bc852e96d2985df9623f
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/2752533
Reviewed-by: Sean Paul <seanpaul@chromium.org>
Reviewed-by: Douglas Anderson <dianders@chromium.org>
Tested-by: Douglas Anderson <dianders@chromium.org>
Commit-Queue: Douglas Anderson <dianders@chromium.org>