CHROMIUM: config: Turn off CONFIG_SCHED_AUTOGROUP on arm64
As far as I can tell it was inadvertently turned on in:

  commit 59c956983063cd940b0bdb7be8f6a527f3cb1c2f
  Author:     Guenter Roeck <groeck@chromium.org>
  AuthorDate: Wed Sep 26 09:41:43 2018 -0700
  Commit:     Guenter Roeck <groeck@chromium.org>
  CommitDate: Tue Oct 23 23:43:18 2018 -0700

      CHROMIUM: Normalize defconfig

That was the defconfig normalization after the rebase to v4.19.
There's no reason in that change about why CONFIG_SCHED_AUTOGROUP was
turned on and no other history can be found.

There's no known reason why we'd want this config on for arm64 but not
for our other platforms and apparently "The autogroups seem to form
many different CGroups which could be slowing down the scheduler since
it iterates through them all in hot paths."

Let's turn it off.  If later we find a reason it needs to be on, we
should probably enable it across the board and make sure that the
overhead is somehow avoided or decide that the overhead is worth it.

This is:
  echo 'CONFIG_SCHED_AUTOGROUP=n' >> chromeos/config/arm64/common.config
  echo 'CONFIG_CGROUPS=y' >> chromeos/config/arm64/common.config
  echo 'CONFIG_CGROUP_SCHED=y' >> chromeos/config/arm64/common.config
  ./chromeos/scripts/kernelconfig olddefconfig
...atop a normalized config.

BUG=b:176927931
TEST=Overhead is gone; system still works normally.

Signed-off-by: Douglas Anderson <dianders@chromium.org>
Change-Id: Id54317fb607ecf5c5fb0ecacd42e55ac8a594ff2
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/2643300
Reviewed-by: Guenter Roeck <groeck@chromium.org>