watchdog: qcom: Remove incorrect usage of QCOM_WDT_ENABLE_IRQ
As per register documentation, QCOM_WDT_ENABLE_IRQ which is BIT(1)
of watchdog control register is wakeup interrupt enable bit and
not related to bark interrupt at all, BIT(0) is used for that.
So remove incorrect usage of this bit when supporting bark irq for
pre-timeout notification. Currently with this bit set and bark
interrupt specified, no pre-timeout notification and/or watchdog
reset/bite occurs.

Fixes: 36375491a439 ("watchdog: qcom: support pre-timeout when the bark irq is available")
Signed-off-by: Sai Prakash Ranjan <saiprakash.ranjan@codeaurora.org>
Change-Id: If39652ab94c5c5616daf265fa279c4399b8522eb