BACKPORT: FROMGIT: futex: Mark the begin of futex exit explicitly
commit 18f694385c4fd77a09851fd301236746ca83f3cb upstream.

Instead of relying on PF_EXITING use an explicit state for the futex exit
and set it in the futex exit function. This moves the smp barrier and the
lock/unlock serialization into the futex code.

As with the DEAD state this is restricted to the exit path as exec
continues to use the same task struct.

This allows to simplify that logic in a next step.

Disallow-Recycled-Builds: test-failures
Signed-off-by: Thomas Gleixner <tglx@linutronix.de>
Reviewed-by: Ingo Molnar <mingo@kernel.org>
Acked-by: Peter Zijlstra (Intel) <peterz@infradead.org>
Link: https://lkml.kernel.org/r/20191106224556.539409004@linutronix.de
Signed-off-by: Greg Kroah-Hartman <gregkh@linuxfoundation.org>
Signed-off-by: Lee Jones <lee.jones@linaro.org>
Signed-off-by: Greg Kroah-Hartman <gregkh@linuxfoundation.org>
(cherry picked from commit 40da511f787b75a80efa3ae5d3f90753a795a9e4
 git://git.kernel.org/pub/scm/linux/kernel/git/stable/linux-stable.git linux-4.4.y)

Conflicts:
   kernel/exit.c
   	context conflict against commit 69991a35e3c22 ("ANDROID: FIXUP:
	sched/tune: fix accounting for runnable tasks")

BUG=b:179467811
TEST=Run CQ

Change-Id: I8881e5f427aec07328fe71adc7f91e96158690c7
Signed-off-by: Guenter Roeck <groeck@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/2688665