soc: qcom: geni: HACK to Fix NULL pointer dereference
pdev struct doesn't exits for the devices whose status are 
disabled from DT node or whose probe has not yet done.
In such cases NULL is returned from 'platform_get_drvdata'.
Next when we try to get driver_data from pdev struct NULL 
pointer dereference is triggered.

Add a NULL check for return values to fix the issue.

We were hitting this issue when one of the QUP probing is not
yet done.

Change-Id: Ibc79f0eb85548d5013d54ae66f99f65d20d3ecb5
Signed-off-by: Roja Rani Yarubandi <rojay@codeaurora.org>