drm/msm: Fix legacy relocs path
In moving code around, we ended up using the same pointer to
copy_from_user() the relocs tables as we used for the cmd table
entry, which is clearly not right.  This went unnoticed because
modern mesa on non-ancent kernels does not actually use relocs.
But this broke ancient mesa on modern kernels.

Change-Id: Ia609a58d8970acbb142a7cde73e034d679ba07d1
Reported-by: Emil Velikov <emil.velikov@collabora.com>
Fixes: 20224d715a88 ("drm/msm/submit: Move copy_from_user ahead of locking bos")
Signed-off-by: Rob Clark <robdclark@chromium.org>
Reviewed-by: Akhil P Oommen <akhilpo@codeaurora.org>
Git-commit: c8d99bb938d3303989c4988caf090084073e85a2
Git-repo: https://gitlab.freedesktop.org/drm/msm.git
Signed-off-by: Akhil P Oommen <akhilpo@codeaurora.org>