[RFC] kvm,x86: Don't preempt VCPU that has preempt_count or IRQs are disabled.
With this CL and the corresponding guest CL, it is possible to use a
non-bound RT VCPU without getting hundreds of milliseconds of latency
spikes when the RT VCPU preempts another VCPU that's holding a lock.

We achieve this by preventing non-RT VCPUs from getting preempted if
they are in a critical section, which we define as having non-zero
preempt_count or interrupts disabled. This avoids the priority inversion
of a RT VCPU preempting another VCPU that is holding a lock that it
wants.

On a machine with 2 CPU threads, creating a VM with 3 VCPUs and making
VCPU 2 RT in the host (and preventing most tasks from running on it by
using "isolcpus"), the max latency shown by
"/data/local/tmp/cyclictest -l100000 -m -Sp90 -i1000 -q -n" in the guest
for VCPU2 never exceeds 4ms (and is often < 1ms) in my tests, even when
the host CPUs are overloaded by running something like
"for i in `seq 8`; do yes > /dev/null & done".
Without these changes, the max latency would often get in the hundreds
of milliseconds.

BUG=b:181909239
TEST=Described above.

Signed-off-by: Suleiman Souhlal <suleiman@google.com>
Change-Id: I120bfa8e528c0ef9c54656fe8d9858a2b7c09783
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/3037941
Commit-Queue: Kazuhiro Inaba <kinaba@chromium.org>
Commit-Queue: Suleiman Souhlal <suleiman@chromium.org>
Tested-by: Kazuhiro Inaba <kinaba@chromium.org>
Auto-Submit: Kazuhiro Inaba <kinaba@chromium.org>
Reviewed-by: Suleiman Souhlal <suleiman@chromium.org>