FIXUP: CHROMIUM: arm64: qcom: sc7180: trogdor: Adjust thermal configuration for big CPUs on pompom
(use phandles to reference thermal zones and trip points, instead of
 replicating the node hierarchy)

Lower the trip point temperatures and sustainable power for the big CPUs
on pompom to meet skin temperature requirements. The sustainable power
value corresponds to a CPU frequency of 1.709 GHz.

This might be an interim solution, using a thermistor to measure/model
skin temperature is being discussed.

BUG=b:171431499
TEST=ODM confirmed that skin temperature requirements are met
  verified that the resulting .dtb file is the same

Change-Id: I04af6f94e2b9e8d26f0abbaebdbf15ef1b114990
Signed-off-by: Matthias Kaehlcke <mka@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/2693764