CHROMIUM: mac80211: drop A-MSDUs on old ciphers
With old ciphers (WEP and TKIP) we shouldn't be using A-MSDUs
since A-MSDUs are only supported if we know that they are, and
the only practical way for that is HT support which doesn't
support old ciphers.

However, we would normally accept them anyway. Since we check
the MMIC before deaggregating A-MSDUs, and the A-MSDU bit in
the QoS header is not protected in TKIP (or WEP), this enables
attacks similar to CVE-2020-24588. To prevent that, drop A-MSDUs
completely with old ciphers.

BUG=b:179418220
TEST=manual testing on M

Signed-off-by: Johannes Berg <johannes.berg@intel.com>
Signed-off-by: vadiraj kakhandaki <vadirajk@google.com>
Change-Id: I01f28af42ee62ab3218d4aa0b3c8195fc5d974ee
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/2921860
Tested-by: ANKIT DARSI AMARNATHAGUPTHA <dankit@google.com>
Reviewed-by: Julan Hsu <julanhsu@google.com>
Reviewed-by: Maharaja Kennadyrajan <mkenna@qualcomm.corp-partner.google.com>
(cherry picked from commit 29df1424cd31449a8dafbfc79c13a6f2d2a47072)
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/2960696
Commit-Queue: Julan Hsu <julanhsu@google.com>