CHROMIUM: timekeeping: Advance CLOCK_BOOTTIME 5 seconds at boot
This is a short-term solution for solving b/162547792. CTS Verifier
expects CLOCK_BOOTTIME and CLOCK_MONOTONIC are different enough to check
if sensors are using the right clock source.

This assumption is not satisfied in the current implementation since VM
guest kernel does not experience suspend/resume cycle and the current
kernel implementation adjusts a gap between CLOCK_BOOTTIME and
CLOCK_MONOTONIC only when the kernel has been resumed from suspension.

To make CTS Verifier happy, we need to adjust CLOCK_BOOTTIME in some way
but there is no viable solution for the real problem yet so we decided
to advance CLOCK_BOOTTIME a bit at boot.

Once the underlying issue is resolved, this patch should be reverted.

BUG=b:162547792
TEST=built and passed Device Suspend Tests in CTS Verifier

Change-Id: Ic9e2bd81de820d77061d3bd6e51787d8e428b630
Signed-off-by: Hikaru Nishida <hikalium@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/2699787
Reviewed-by: Suleiman Souhlal <suleiman@chromium.org>