INTERNAL: coresight: add support to enable multiple coresight paths
Current coresight implementation supports enabling of a
single path. This patch adds support to enable multiple source
to sink paths. We build a path from source to sink when any
source is enabled and store it in a list. When the source is
disabled, we fetch the corresponding path from the list and
decrement the refcount on each device in the path. The device
is disabled if the refcount reaches zero. Don't store path
to coresight data structure of source to avoid unnecessary
change to ABI.

Change-Id: Ic90e11067907425fff92b2c88d31e4c5d6aff910
Signed-off-by: Tingwei Zhang <tingwei@codeaurora.org>
Signed-off-by: Satyajit Desai <sadesai@codeaurora.org>
Signed-off-by: Rama Aparna Mallavarapu <aparnam@codeaurora.org>
Signed-off-by: Mulu He <muluhe@codeaurora.org>
Signed-off-by: Sai Prakash Ranjan <saiprakash.ranjan@codeaurora.org>