CHROMIUM: cfg80211: mitigate A-MSDU aggregation attacks
Mitigate A-MSDU injection attacks (CVE-2020-24588) by detecting if the
destination address of a subframe equals an RFC1042 (i.e., LLC/SNAP)
header, and if so dropping the complete A-MSDU frame. This mitigates
known attacks, although new (unknown) aggregation-based attacks may
remain possible.

This defense works because in A-MSDU aggregation injection attacks, a
normal encrypted Wi-Fi frame is turned into an A-MSDU frame. This means
the first 6 bytes of the first A-MSDU subframe correspond to an RFC1042
header. In other words, the destination MAC address of the first A-MSDU
subframe contains the start of an RFC1042 header during an aggregation
attack. We can detect this and thereby prevent this specific attack.
For details, see Section 7.2 of "Fragment and Forge: Breaking Wi-Fi
Through Frame Aggregation and Fragmentation".

Note that for kernel 4.9 and above this patch depends on "mac80211:
properly handle A-MSDUs that start with a rfc1042 header". Otherwise
this patch has no impact and attacks will remain possible.

(cherry picked from commit e4e86e6adf6e4d88adf3e6e755f57e86aec6a9f2)

BUG=b:179418147
TEST=manual testing on G

Signed-off-by: Mathy Vanhoef <Mathy.Vanhoef@kuleuven.be>
Signed-off-by: Johannes Berg <johannes.berg@intel.com>
Signed-off-by: Luca Coelho <luciano.coelho@intel.com>
Signed-off-by: Vadiraj Kakhandaki <vadirajk@google.com>
Change-Id: Ia75390fea58933e319288456ce1f52d408417cdb
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/2967268
Tested-by: ANKIT DARSI AMARNATHAGUPTHA <dankit@google.com>
Reviewed-by: Julan Hsu <julanhsu@google.com>
Commit-Queue: Julan Hsu <julanhsu@google.com>