CHROMIUM: mac80211: drop A-MSDUs on old ciphers
With old ciphers (WEP and TKIP) we shouldn't be using A-MSDUs
since A-MSDUs are only supported if we know that they are, and
the only practical way for that is HT support which doesn't
support old ciphers.

However, we would normally accept them anyway. Since we check
the MMIC before deaggregating A-MSDUs, and the A-MSDU bit in
the QoS header is not protected in TKIP (or WEP), this enables
attacks similar to CVE-2020-24588. To prevent that, drop A-MSDUs
completely with old ciphers.

BUG=b:179418147
TEST=manual testing on G

Signed-off-by: Johannes Berg <johannes.berg@intel.com>
Signed-off-by: vadiraj kakhandaki <vadirajk@google.com>
Change-Id: Id1dfe83d77a06fe2a1be7dabf14a1156f4d08148
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/2914917
Reviewed-by: Maharaja Kennadyrajan <mkenna@qualcomm.corp-partner.google.com>
Reviewed-by: Julan Hsu <julanhsu@google.com>
(cherry picked from commit 5d6130298f4409f1c9d316a34bf2eb35db125af7)
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/2958981
Reviewed-by: ANKIT DARSI AMARNATHAGUPTHA <dankit@google.com>
Tested-by: ANKIT DARSI AMARNATHAGUPTHA <dankit@google.com>
Commit-Queue: Julan Hsu <julanhsu@google.com>