CHROMIUM: mac80211: drop A-MSDUs on old ciphers
With old ciphers (WEP and TKIP) we shouldn't be using A-MSDUs
since A-MSDUs are only supported if we know that they are, and
the only practical way for that is HT support which doesn't
support old ciphers.

However, we would normally accept them anyway. Since we check
the MMIC before deaggregating A-MSDUs, and the A-MSDU bit in
the QoS header is not protected in TKIP (or WEP), this enables
attacks similar to CVE-2020-24588. To prevent that, drop A-MSDUs
completely with old ciphers.

(cherry picked from commit 5d6130298f4409f1c9d316a34bf2eb35db125af7)

BUG=b:179418147
TEST=manual testing on G

Signed-off-by: Johannes Berg <johannes.berg@intel.com>
Signed-off-by: Vadiraj Kakhandaki <vadirajk@google.com>
Change-Id: If8f557b5edb868f489fb4025e449d5dac53471cd
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/2967269
Tested-by: ANKIT DARSI AMARNATHAGUPTHA <dankit@google.com>
Reviewed-by: Julan Hsu <julanhsu@google.com>
Commit-Queue: Julan Hsu <julanhsu@google.com>