CHROMIUM: mac80211: do not accept/forward invalid EAPOL frames
EAPOL frames are used for authentication and key management between the
AP and each individual STA associated in the BSS. Those frames are not
supposed to be sent by one associated STA to another associated STA
(either unicast for broadcast/multicast).

Similarly, in 802.11 they're supposed to be sent to the authenticator
(AP) address.

Since it is possible for unexpected EAPOL frames to result in misbehavior
in supplicant implementations, it is better for the AP to not allow such
cases to be forwarded to other clients either directly, or indirectly if
the AP interface is part of a bridge.

Accept EAPOL (control port) frames only if they're transmitted to the
own address, or, due to interoperability concerns, to the PAE group
address.

Disable forwarding of EAPOL (or well, the configured control port
protocol) frames back to wireless medium in all cases. Previously, these
frames were accepted from fully authenticated and authorized stations
and also from unauthenticated stations for one of the cases.

Additionally, to avoid forwarding by the bridge, rewrite the PAE group
address case to the local MAC address.

(cherry picked from commit 6ec18700c9e331aa36afff4996b35c4e83a8375a)

BUG=b:179418220
TEST=manual testing on M

Co-developed-by: Jouni Malinen <jouni@codeaurora.org>
Signed-off-by: Jouni Malinen <jouni@codeaurora.org>
Signed-off-by: Johannes Berg <johannes.berg@intel.com>
Signed-off-by: Vadiraj Kakhandaki <vadirajk@google.com>
Change-Id: I1fc816e4f5e52f9a567734daa352181ebff07218
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/2967538
Tested-by: ANKIT DARSI AMARNATHAGUPTHA <dankit@google.com>
Reviewed-by: Julan Hsu <julanhsu@google.com>
Commit-Queue: Julan Hsu <julanhsu@google.com>