FROMLIST: sched: Prepare for Core-wide rq->lock
When switching on core-sched, CPUs need to agree which lock to use for
their RQ.

The new rule will be that rq->core_enabled will be toggled while
holding all rq->__locks that belong to a core. This means we need to
double check the rq->core_enabled value after each lock acquire and
retry if it changed.

This also has implications for those sites that take multiple RQ
locks, they need to be careful that the second lock doesn't end up
being the first lock.

Verify the lock pointer after acquiring the first lock, because if
they're on the same core, holding any of the rq->__lock instances will
pin the core state.

While there, change the rq->__lock order to CPU number, instead of rq
address, this greatly simplifies the next patch.

BUG=b:152605392
TEST=run power_VideoCall test

(am from
https://lore.kernel.org/lkml/20210422123308.196692074@infradead.org/)

Signed-off-by: Peter Zijlstra (Intel) <peterz@infradead.org>
Tested-by: Don Hiatt <dhiatt@digitalocean.com>
Signed-off-by: Joel Fernandes <joelaf@google.com>
Change-Id: Iaae2b590f5130a6cba8ca75c10f0a558e1f60175
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/2880779
Reviewed-by: Sonny Rao <sonnyrao@chromium.org>