CHROMIUM: drm/i915/lspcon: Increase LSPCON timeout
400ms is not enough time for the Parade PS175 LSPCON adapter on
Ambassador to settle. This is causing blank screens on boot. Empirical
testing can no longer reproduce the error with an increase to 425ms, so
pad to 500ms to give some margin of error.

We are increasing the timeout while we work on root-causing why 400ms is
not enough time to settle.

BUG=b:178169843
TEST=build kernel; deploy to device; reboot 100x

Change-Id: I259f2aaeba3b81d025c01d15bfde02b9564476f5
Signed-off-by: Matt Ziegelbaum <ziegs@google.com>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/2727925
Reviewed-by: Joe Tessler <jrt@chromium.org>
Commit-Queue: Matthew Ziegelbaum <ziegs@chromium.org>
Tested-by: Matthew Ziegelbaum <ziegs@chromium.org>