soc: qcom: rpmhpd: sc7280: Add support for mx max corner vote
On sc7280, in order to reliably blow fuses, we need to push mx to
a corner higher than whats supported for any other consumer on the SoC.
In order to avoid exposing this via command DB and letting
some conumers who vote with a INT_MAX to wrongly push mx to
a level much higher than whats needed, handle this with a magix hlvl
used only by the qfprom consumer, and send the max supported vlvl
value to rpmh.

The reason this higher corner is not safe in a general scenario is due
to a possibility of peak mx current at this corner, causing an abnormal
PMIC reset. Given fuse blowing is done in a more controlled environment,
in factory, with very little to no concurrent usercases active, this
corner is considered safe for the fuse blow operation alone.

Signed-off-by: Rajendra Nayak <rnayak@codeaurora.org>
Change-Id: Ie05608e47331ecda5dc230b3234155daf0415d7a