INTERNAL: coresight: Add symbolic links for coresight device nodes
Add symbolic links to coresight device nodes so that userspace components
such as diag-router or some daemons which relies on older coresight device
names which are passed from DT property still continue to work.

For ex: Currently for ETM timestamp tests, the tmc etr and etf nodes
used by test team scripts are:
  /sys/bus/coresight/devices/tmc_etr0/
  /sys/bus/coresight/devices/tmc_etf0/

And for diag over qdss use cases, userspace scripts use this path,
  /sys/bus/coresight/devices/coresight-tmc-etr/
  /sys/bus/coresight/devices/coresight-tmc-etf/

So in order to maintain both and not break either, we create a
symlink such that both path exists and are same.

Change-Id: I1b694df0149fac8094fc0c0218bb38a15645a588
Signed-off-by: Sai Prakash Ranjan <saiprakash.ranjan@codeaurora.org>