CHROMIUM: media: camx: Change most "select" to "depends on"
Many of the features that the camx Kconfig was "select"ing are
features that a user normally has control over (AKA they show up in
menuconfig). It's generally not ideal to "select" such features and
"select" is usually only used for auto-enabling hunks of code that are
shared between drivers but not something users care about. Probably
when camera code is added upstream they will want the "select" changed
to a "depends on".

From a more practical point of view, the fact that camx was
"select"ing things was causing a bunch of things to fall out of our
"defconfig" as can be seen in CL:2735734. That has no effect on
functionality but could cause later problems. Specifically, the
following would be bad:

1. Land camx CL that does "select INTERCONNECT_QCOM_SC7180"
2. Normalize defconfig and "INTERCONNECT_QCOM_SC7180" no longer shows
   up in defconfig (since it's implied by another config).
3. We replace camx with a newer version that _doesn't_ do select but
   does "depends on".
4. "INTERCONNECT_QCOM_SC7180" is no longer implied and also no longer
   explicitly specified so it gets disabled.
5. Now someone has to debug how the "INTERCONNECT_QCOM_SC7180" got
   disabled.

Let's just fix it now.

NOTE: this has a net effect that if we re-land CL:2352970 we will no
longer enable the i2c driver CONFIG_I2C_QUP nor will we enable
CONFIG_QTI_CCI. As talked about elsewhere the i2c driver from
CONFIG_I2C_QUP is only needed for older (pre-geni) Qualcomm SoCs. The
fact that the CCI driver is tied to it means one of these things:
a) We don't need the CCI driver.
b) The CCI driver didn't actually have any dependency on the i2c
   driver and we should simply remove the dependency.
c) The CCI driver was needed and did have a dependency on the old i2c
   driver. In this case the CCI driver should be updated to remove
   this dependency.

BUG=b:168139312, b:181965641
TEST=Re-pick CL:2352970 and use "genconfig" to compare before/after.

Disallow-Recycled-Builds: test-failures
Signed-off-by: Douglas Anderson <dianders@chromium.org>
Signed-off-by: Guenter Roeck <groeck@chromium.org>
Change-Id: Ibff7aee45b61e276cef3750f3606b8f71aabffc0
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/2747261
(cherry picked from commit 6edc103697f577da1916c903ae2932264e6bd512)
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/2799574
Reviewed-by: Tomasz Figa <tfiga@chromium.org>
Reviewed-by: Jian-Jia Su <jjsu@chromium.org>
Commit-Queue: Jian-Jia Su <jjsu@chromium.org>
Tested-by: Jian-Jia Su <jjsu@chromium.org>