CHROMIUM: iwl7000: mac80211: don't apply flow control on management frames
In some cases (depending on the driver, but it's true e.g. for
iwlwifi) we're using an internal TXQ for management packets,
mostly to simplify the code and to have a place to queue them.
However, it appears that in certain cases we can confuse the
code and management frames are dropped, which is certainly not
what we want.

Short-circuit the processing of management frames. To keep the
impact minimal, only put them on the frags queue and check the
tid == management only for doing that and to skip the airtime
fairness checks, if applicable.

BUG=b:175364711
TEST=TBD

Change-Id: Id2b66eb5a17f3869b776c39b5ca713272ea09d5d
Signed-off-by: Johannes Berg <johannes.berg@intel.com>
(cherry picked from commit 35c058f0f0b8dfd1bf42a371daf8d8af80cd4490)
iwl7000-tree: 140e7cd7fc02b6c69cb7c8f5fb0954befe137fed
Signed-off-by: Golan Ben Ami <golan.ben.ami@intel.com>
Signed-off-by: Luca Coelho <luciano.coelho@intel.com>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/2819154
Tested-by: Allen Cheng <allen.cheng@lcfc.corp-partner.google.com>
Reviewed-by: Nicolas Norvez <norvez@chromium.org>
Commit-Queue: Will Lee <wwlee@google.com>