CHROMIUM: ASoC: rt5682: read property to get mclk control
The systems can use ACPI DSDT to pass on the mclk to rt5682.
Enable/Disable of the mclk is already handled in the codec so
platform drivers don't have to explicitly do handling of mclk.

Note:
This patch follows the patch of dialog's codec driver to
do the same thing with the AMD platform.
https://chromium.googlesource.com/chromiumos/third_party/kernel/+/71e6a8a729e74786dfeb01b5c7e9e1a56ac484dc%5E%21/#F1
However, this patch is not a typical approach to get clock control.
This patch will not be acceptable to ALSA maintainers.
Some comments from the maintainer:
https://www.mail-archive.com/search?l=linux-kernel@vger.kernel.org&q=subject:%22Re%5C%3A+%5C%5BPATCH+v3%5C%5D+ASoC%5C%3A+da7219%5C%3A+read+fmw+property+to+get+mclk+for+non%5C-dts+systems%22&o=newest&f=1
https://mailman.alsa-project.org/pipermail/alsa-devel/2021-March/182568.html
Therefore, this patch will be Chromium change.

BUG=b:171755306
TEST=tested by OEM/AMD on Barla/Mordin

Signed-off-by: Shuming Fan <shumingf@realtek.com>
Change-Id: I89d22786fb92cef9289dec52b4eab2d96f3c4edd
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/2767117
Reviewed-by: Yu-Hsuan Hsu <yuhsuan@chromium.org>
Commit-Queue: Yu-Hsuan Hsu <yuhsuan@chromium.org>
Tested-by: Yu-Hsuan Hsu <yuhsuan@chromium.org>
Signed-off-by: derek.fang <derek.fang@realtek.corp-partner.google.com>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/2914870