UPSTREAM: selinux: vsock: Set SID for socket returned by accept()
aosp/1645810 has been merged to kernel android12-5.10
We need this for ARCVM as well (b/149888451)

For AF_VSOCK, accept() currently returns sockets that are unlabelled.
Other socket families derive the child's SID from the SID of the parent
and the SID of the incoming packet. This is typically done as the
connected socket is placed in the queue that accept() removes from.

Reuse the existing 'security_sk_clone' hook to copy the SID from the
parent (server) socket to the child. There is no packet SID in this
case.

Fixes: d021c344051a ("VSOCK: Introduce VM Sockets")
Signed-off-by: David Brazdil <dbrazdil@google.com>
Signed-off-by: David S. Miller <davem@davemloft.net>
Signed-off-by: Satoshi Niwa <niwa@chromium.org>
(cherry picked from commit 1f935e8e72ec28dddb2dc0650b3b6626a293d94b
 https://git.kernel.org/pub/scm/linux/kernel/git/netdev/net.git master)
Link: https://lore.kernel.org/r/20210319130541.2188184-1-dbrazdil@google.com
BUG=b:130668487
BUG=b:149888451
TEST=arc.Boot.vm
Change-Id: Idb656917cc600df7cb83cb15a6c0f0499045d77f
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/2796853
Reviewed-by: Sean Paul <seanpaul@chromium.org>
Reviewed-by: Jeffrey Vander Stoep <jeffv@google.com>
Reviewed-by: Hikaru Nishida <hikalium@chromium.org>