FROMLIST: platform/x86: intel_pmc_core: Remove global struct pmc_dev
The intel_pmc_core driver did not always bind to a device which meant it
lacked a struct device that could be used to maintain driver data. So a
global instance of struct pmc_dev was used for this purpose and functions
accessed this directly. Since the driver now binds to an ACPI device,
remove the global pmc_dev in favor of one that is allocated during probe.
Modify users of the global to obtain the object by argument instead.

Signed-off-by: David E. Box <david.e.box@linux.intel.com>
(am from https://lore.kernel.org/patchwork/patch/1406066/)

BUG=b:179699891
TEST=check pmc_core debug interface presence
TEST=check slp0 residency values on ADL-P

Signed-off-by: Nathan Ciobanu <nathan.d.ciobanu@intel.com>
Change-Id: I7a78107b6fee6887a913e0845b053048712f1d82
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/2800277
Reviewed-by: Sean Paul <seanpaul@chromium.org>
Reviewed-by: Alex Levin <levinale@google.com>