drm/msm/disp/dpu1: increase range of dpu_irq_map
Currently, each register in the dpu interrupt set is allowed
to have a maximum of 32 interrupts. With the introduction
of INTF_5_VSYNC and INTF_5_UNDERRUN irqs for EDP panel,
the total number of interrupts under INTR_STATUS register
in dpu_irq_map will exceed 32. Increase the range of each
interrupt register to 64 to handle this.

Change-Id: Ibcc2434079f49c6327db162c7a5c2c7fa23e907c
Signed-off-by: Krishna Manikandan <mkrishn@codeaurora.org>